requirements.txt
playwright>=1.48.0
pandas>=2.2.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0httpx>=0.27.0